@@ -4,7 +4,6 @@
 # Use of this source code is governed by a BSD-style license that can be
 # found in the LICENSE file.
 
-from __future__ import print_function
 
 import copy
 import gyp.input
@@ -193,7 +192,7 @@ def ShlexEnv(env_name):
 
 def FormatOpt(opt, value):
     if opt.startswith("--"):
-        return "%s=%s" % (opt, value)
+        return f"{opt}={value}"
     return opt + value
 
 
