@@ -10,17 +10,12 @@ import tempfile
 import sys
 import subprocess
 
-try:
-    from collections.abc import MutableSet
-except ImportError:
-    from collections import MutableSet
-
-PY3 = bytes != str
+from collections.abc import MutableSet
 
 
 # A minimal memoizing decorator. It'll blow up if the args aren't immutable,
 # among other "problems".
-class memoize(object):
+class memoize:
     def __init__(self, func):
         self.func = func
         self.cache = {}
@@ -348,7 +343,7 @@ def WriteOnDiff(filename):
     the target if it differs (on close).
   """
 
-    class Writer(object):
+    class Writer:
         """Wrapper around file which only covers the target if it differs."""
 
         def __init__(self):
@@ -566,8 +561,8 @@ class OrderedSet(MutableSet):
 
     def __repr__(self):
         if not self:
-            return "%s()" % (self.__class__.__name__,)
-        return "%s(%r)" % (self.__class__.__name__, list(self))
+            return f"{self.__class__.__name__}()"
+        return "{}({!r})".format(self.__class__.__name__, list(self))
 
     def __eq__(self, other):
         if isinstance(other, OrderedSet):
@@ -653,9 +648,7 @@ def IsCygwin():
         out = subprocess.Popen(
             "uname", stdout=subprocess.PIPE, stderr=subprocess.STDOUT
         )
-        stdout, stderr = out.communicate()
-        if PY3:
-            stdout = stdout.decode("utf-8")
+        stdout = out.communicate()[0].decode("utf-8")
         return "CYGWIN" in str(stdout)
     except Exception:
         return False
