@@ -84,7 +84,7 @@ def _ConstructContentList(xml_parts, specification, pretty, level=0):
     rest = specification[1:]
     if rest and isinstance(rest[0], dict):
         for at, val in sorted(rest[0].items()):
-            xml_parts.append(' %s="%s"' % (at, _XmlEscape(val, attr=True)))
+            xml_parts.append(' {}="{}"'.format(at, _XmlEscape(val, attr=True)))
         rest = rest[1:]
     if rest:
         xml_parts.append(">")
@@ -101,7 +101,7 @@ def _ConstructContentList(xml_parts, specification, pretty, level=0):
                 _ConstructContentList(xml_parts, child_spec, pretty, level + 1)
         if multi_line and indentation:
             xml_parts.append(indentation)
-        xml_parts.append("</%s>%s" % (name, new_line))
+        xml_parts.append(f"</{name}>{new_line}")
     else:
         xml_parts.append("/>%s" % new_line)
 
@@ -125,9 +125,9 @@ def WriteXmlIfChanged(content, path, encoding="utf-8", pretty=False, win32=False
 
     # Get the old content
     try:
-        with open(path, "r") as file:
+        with open(path) as file:
             existing = file.read()
-    except IOError:
+    except OSError:
         existing = None
 
     # It has changed, write it
