@@ -14,7 +14,6 @@
 # variables set potentially clash with other Android build system variables.
 # Try to avoid setting global variables where possible.
 
-from __future__ import print_function
 
 import gyp
 import gyp.common
@@ -84,7 +83,7 @@ def IsCPPExtension(ext):
 
 def Sourceify(path):
     """Convert a path to its source directory form. The Android backend does not
-     support options.generator_output, so this function is a noop."""
+    support options.generator_output, so this function is a noop."""
     return path
 
 
@@ -100,11 +99,11 @@ target_outputs = {}
 target_link_deps = {}
 
 
-class AndroidMkWriter(object):
+class AndroidMkWriter:
     """AndroidMkWriter packages up the writing of one target-specific Android.mk.
 
-  Its only real entry point is Write(), and is mostly used for namespacing.
-  """
+    Its only real entry point is Write(), and is mostly used for namespacing.
+    """
 
     def __init__(self, android_top_dir):
         self.android_top_dir = android_top_dir
@@ -123,18 +122,18 @@ class AndroidMkWriter(object):
     ):
         """The main entry point: writes a .mk file for a single target.
 
-    Arguments:
-      qualified_target: target we're generating
-      relative_target: qualified target name relative to the root
-      base_path: path relative to source root we're building in, used to resolve
-                 target-relative paths
-      output_filename: output .mk file name to write
-      spec, configs: gyp info
-      part_of_all: flag indicating this target is part of 'all'
-      write_alias_target: flag indicating whether to create short aliases for
-                          this target
-      sdk_version: what to emit for LOCAL_SDK_VERSION in output
-    """
+        Arguments:
+          qualified_target: target we're generating
+          relative_target: qualified target name relative to the root
+          base_path: path relative to source root we're building in, used to resolve
+                     target-relative paths
+          output_filename: output .mk file name to write
+          spec, configs: gyp info
+          part_of_all: flag indicating this target is part of 'all'
+          write_alias_target: flag indicating whether to create short aliases for
+                              this target
+          sdk_version: what to emit for LOCAL_SDK_VERSION in output
+        """
         gyp.common.EnsureDirExists(output_filename)
 
         self.fp = open(output_filename, "w")
@@ -254,15 +253,15 @@ class AndroidMkWriter(object):
     def WriteActions(self, actions, extra_sources, extra_outputs):
         """Write Makefile code for any 'actions' from the gyp input.
 
-    extra_sources: a list that will be filled in with newly generated source
-                   files, if any
-    extra_outputs: a list that will be filled in with any outputs of these
-                   actions (used to make other pieces dependent on these
-                   actions)
-    """
+        extra_sources: a list that will be filled in with newly generated source
+                       files, if any
+        extra_outputs: a list that will be filled in with any outputs of these
+                       actions (used to make other pieces dependent on these
+                       actions)
+        """
         for action in actions:
             name = make.StringToMakefileVariable(
-                "%s_%s" % (self.relative_target, action["action_name"])
+                "{}_{}".format(self.relative_target, action["action_name"])
             )
             self.WriteLn('### Rules for action "%s":' % action["action_name"])
             inputs = action["inputs"]
@@ -350,7 +349,7 @@ class AndroidMkWriter(object):
             for output in outputs[1:]:
                 # Make each output depend on the main output, with an empty command
                 # to force make to notice that the mtime has changed.
-                self.WriteLn("%s: %s ;" % (self.LocalPathify(output), main_output))
+                self.WriteLn("{}: {} ;".format(self.LocalPathify(output), main_output))
 
             extra_outputs += outputs
             self.WriteLn()
@@ -360,11 +359,11 @@ class AndroidMkWriter(object):
     def WriteRules(self, rules, extra_sources, extra_outputs):
         """Write Makefile code for any 'rules' from the gyp input.
 
-    extra_sources: a list that will be filled in with newly generated source
-                   files, if any
-    extra_outputs: a list that will be filled in with any outputs of these
-                   rules (used to make other pieces dependent on these rules)
-    """
+        extra_sources: a list that will be filled in with newly generated source
+                       files, if any
+        extra_outputs: a list that will be filled in with any outputs of these
+                       rules (used to make other pieces dependent on these rules)
+        """
         if len(rules) == 0:
             return
 
@@ -372,7 +371,7 @@ class AndroidMkWriter(object):
             if len(rule.get("rule_sources", [])) == 0:
                 continue
             name = make.StringToMakefileVariable(
-                "%s_%s" % (self.relative_target, rule["rule_name"])
+                "{}_{}".format(self.relative_target, rule["rule_name"])
             )
             self.WriteLn('\n### Generated for rule "%s":' % name)
             self.WriteLn('# "%s":' % rule)
@@ -452,7 +451,7 @@ class AndroidMkWriter(object):
                 for output in outputs[1:]:
                     # Make each output depend on the main output, with an empty command
                     # to force make to notice that the mtime has changed.
-                    self.WriteLn("%s: %s ;" % (output, main_output))
+                    self.WriteLn(f"{output}: {main_output} ;")
                 self.WriteLn()
 
         self.WriteLn()
@@ -460,9 +459,9 @@ class AndroidMkWriter(object):
     def WriteCopies(self, copies, extra_outputs):
         """Write Makefile code for any 'copies' from the gyp input.
 
-    extra_outputs: a list that will be filled in with any outputs of this action
-                   (used to make other pieces dependent on this action)
-    """
+        extra_outputs: a list that will be filled in with any outputs of this action
+                       (used to make other pieces dependent on this action)
+        """
         self.WriteLn("### Generated for copy rule.")
 
         variable = make.StringToMakefileVariable(self.relative_target + "_copies")
@@ -487,25 +486,25 @@ class AndroidMkWriter(object):
                     self.LocalPathify(os.path.join(copy["destination"], filename))
                 )
 
-                self.WriteLn(
-                    "%s: %s $(GYP_TARGET_DEPENDENCIES) | $(ACP)" % (output, path)
-                )
+                self.WriteLn(f"{output}: {path} $(GYP_TARGET_DEPENDENCIES) | $(ACP)")
                 self.WriteLn("\t@echo Copying: $@")
                 self.WriteLn("\t$(hide) mkdir -p $(dir $@)")
                 self.WriteLn("\t$(hide) $(ACP) -rpf $< $@")
                 self.WriteLn()
                 outputs.append(output)
-        self.WriteLn("%s = %s" % (variable, " ".join(map(make.QuoteSpaces, outputs))))
+        self.WriteLn(
+            "{} = {}".format(variable, " ".join(map(make.QuoteSpaces, outputs)))
+        )
         extra_outputs.append("$(%s)" % variable)
         self.WriteLn()
 
     def WriteSourceFlags(self, spec, configs):
         """Write out the flags and include paths used to compile source files for
-    the current target.
+        the current target.
 
-    Args:
-      spec, configs: input from gyp.
-    """
+        Args:
+          spec, configs: input from gyp.
+        """
         for configname, config in sorted(configs.items()):
             extracted_includes = []
 
@@ -554,16 +553,16 @@ class AndroidMkWriter(object):
 
     def WriteSources(self, spec, configs, extra_sources):
         """Write Makefile code for any 'sources' from the gyp input.
-    These are source files necessary to build the current target.
-    We need to handle shared_intermediate directory source files as
-    a special case by copying them to the intermediate directory and
-    treating them as a generated sources. Otherwise the Android build
-    rules won't pick them up.
+        These are source files necessary to build the current target.
+        We need to handle shared_intermediate directory source files as
+        a special case by copying them to the intermediate directory and
+        treating them as a generated sources. Otherwise the Android build
+        rules won't pick them up.
 
-    Args:
-      spec, configs: input from gyp.
-      extra_sources: Sources generated from Actions or Rules.
-    """
+        Args:
+          spec, configs: input from gyp.
+          extra_sources: Sources generated from Actions or Rules.
+        """
         sources = filter(make.Compilable, spec.get("sources", []))
         generated_not_sources = [x for x in extra_sources if not make.Compilable(x)]
         extra_sources = filter(make.Compilable, extra_sources)
@@ -617,7 +616,7 @@ class AndroidMkWriter(object):
             if IsCPPExtension(ext) and ext != local_cpp_extension:
                 local_file = root + local_cpp_extension
             if local_file != source:
-                self.WriteLn("%s: %s" % (local_file, self.LocalPathify(source)))
+                self.WriteLn("{}: {}".format(local_file, self.LocalPathify(source)))
                 self.WriteLn("\tmkdir -p $(@D); cp $< $@")
                 origin_src_dirs.append(os.path.dirname(source))
             final_generated_sources.append(local_file)
@@ -640,10 +639,10 @@ class AndroidMkWriter(object):
     def ComputeAndroidModule(self, spec):
         """Return the Android module name used for a gyp spec.
 
-    We use the complete qualified target name to avoid collisions between
-    duplicate targets in different directories. We also add a suffix to
-    distinguish gyp-generated module names.
-    """
+        We use the complete qualified target name to avoid collisions between
+        duplicate targets in different directories. We also add a suffix to
+        distinguish gyp-generated module names.
+        """
 
         if int(spec.get("android_unmangled_name", 0)):
             assert self.type != "shared_library" or self.target.startswith("lib")
@@ -662,7 +661,7 @@ class AndroidMkWriter(object):
             suffix = "_gyp"
 
         if self.path:
-            middle = make.StringToMakefileVariable("%s_%s" % (self.path, self.target))
+            middle = make.StringToMakefileVariable(f"{self.path}_{self.target}")
         else:
             middle = make.StringToMakefileVariable(self.target)
 
@@ -671,11 +670,11 @@ class AndroidMkWriter(object):
     def ComputeOutputParts(self, spec):
         """Return the 'output basename' of a gyp spec, split into filename + ext.
 
-    Android libraries must be named the same thing as their module name,
-    otherwise the linker can't find them, so product_name and so on must be
-    ignored if we are building a library, and the "lib" prepending is
-    not done for Android.
-    """
+        Android libraries must be named the same thing as their module name,
+        otherwise the linker can't find them, so product_name and so on must be
+        ignored if we are building a library, and the "lib" prepending is
+        not done for Android.
+        """
         assert self.type != "loadable_module"  # TODO: not supported?
 
         target = spec["target_name"]
@@ -711,17 +710,17 @@ class AndroidMkWriter(object):
     def ComputeOutputBasename(self, spec):
         """Return the 'output basename' of a gyp spec.
 
-    E.g., the loadable module 'foobar' in directory 'baz' will produce
-      'libfoobar.so'
-    """
+        E.g., the loadable module 'foobar' in directory 'baz' will produce
+          'libfoobar.so'
+        """
         return "".join(self.ComputeOutputParts(spec))
 
     def ComputeOutput(self, spec):
         """Return the 'output' (full output path) of a gyp spec.
 
-    E.g., the loadable module 'foobar' in directory 'baz' will produce
-      '$(obj)/baz/libfoobar.so'
-    """
+        E.g., the loadable module 'foobar' in directory 'baz' will produce
+          '$(obj)/baz/libfoobar.so'
+        """
         if self.type == "executable":
             # We install host executables into shared_intermediate_dir so they can be
             # run by gyp rules that refer to PRODUCT_DIR.
@@ -740,7 +739,7 @@ class AndroidMkWriter(object):
                     % (self.android_class, self.android_module)
                 )
             else:
-                path = "$(call intermediates-dir-for,%s,%s,,,$(GYP_VAR_PREFIX))" % (
+                path = "$(call intermediates-dir-for,{},{},,,$(GYP_VAR_PREFIX))".format(
                     self.android_class,
                     self.android_module,
                 )
@@ -749,14 +748,14 @@ class AndroidMkWriter(object):
         return os.path.join(path, self.ComputeOutputBasename(spec))
 
     def NormalizeIncludePaths(self, include_paths):
-        """ Normalize include_paths.
-    Convert absolute paths to relative to the Android top directory.
+        """Normalize include_paths.
+        Convert absolute paths to relative to the Android top directory.
 
-    Args:
-      include_paths: A list of unprocessed include paths.
-    Returns:
-      A list of normalized include paths.
-    """
+        Args:
+          include_paths: A list of unprocessed include paths.
+        Returns:
+          A list of normalized include paths.
+        """
         normalized = []
         for path in include_paths:
             if path[0] == "/":
@@ -767,11 +766,11 @@ class AndroidMkWriter(object):
     def ExtractIncludesFromCFlags(self, cflags):
         """Extract includes "-I..." out from cflags
 
-    Args:
-      cflags: A list of compiler flags, which may be mixed with "-I.."
-    Returns:
-      A tuple of lists: (clean_clfags, include_paths). "-I.." is trimmed.
-    """
+        Args:
+          cflags: A list of compiler flags, which may be mixed with "-I.."
+        Returns:
+          A tuple of lists: (clean_clfags, include_paths). "-I.." is trimmed.
+        """
         clean_cflags = []
         include_paths = []
         for flag in cflags:
@@ -785,14 +784,14 @@ class AndroidMkWriter(object):
     def FilterLibraries(self, libraries):
         """Filter the 'libraries' key to separate things that shouldn't be ldflags.
 
-    Library entries that look like filenames should be converted to android
-    module names instead of being passed to the linker as flags.
+        Library entries that look like filenames should be converted to android
+        module names instead of being passed to the linker as flags.
 
-    Args:
-      libraries: the value of spec.get('libraries')
-    Returns:
-      A tuple (static_lib_modules, dynamic_lib_modules, ldflags)
-    """
+        Args:
+          libraries: the value of spec.get('libraries')
+        Returns:
+          A tuple (static_lib_modules, dynamic_lib_modules, ldflags)
+        """
         static_lib_modules = []
         dynamic_lib_modules = []
         ldflags = []
@@ -823,10 +822,10 @@ class AndroidMkWriter(object):
     def ComputeDeps(self, spec):
         """Compute the dependencies of a gyp spec.
 
-    Returns a tuple (deps, link_deps), where each is a list of
-    filenames that will need to be put in front of make for either
-    building (deps) or linking (link_deps).
-    """
+        Returns a tuple (deps, link_deps), where each is a list of
+        filenames that will need to be put in front of make for either
+        building (deps) or linking (link_deps).
+        """
         deps = []
         link_deps = []
         if "dependencies" in spec:
@@ -846,9 +845,9 @@ class AndroidMkWriter(object):
     def WriteTargetFlags(self, spec, configs, link_deps):
         """Write Makefile code to specify the link flags and library dependencies.
 
-    spec, configs: input from gyp.
-    link_deps: link dependency list; see ComputeDeps()
-    """
+        spec, configs: input from gyp.
+        link_deps: link dependency list; see ComputeDeps()
+        """
         # Libraries (i.e. -lfoo)
         # These must be included even for static libraries as some of them provide
         # implicit include paths through the build system.
@@ -891,12 +890,12 @@ class AndroidMkWriter(object):
     ):
         """Write Makefile code to produce the final target of the gyp spec.
 
-    spec, configs: input from gyp.
-    deps, link_deps: dependency lists; see ComputeDeps()
-    part_of_all: flag indicating this target is part of 'all'
-    write_alias_target: flag indicating whether to create short aliases for this
-                        target
-    """
+        spec, configs: input from gyp.
+        deps, link_deps: dependency lists; see ComputeDeps()
+        part_of_all: flag indicating this target is part of 'all'
+        write_alias_target: flag indicating whether to create short aliases for this
+                            target
+        """
         self.WriteLn("### Rules for final target.")
 
         if self.type != "none":
@@ -909,7 +908,7 @@ class AndroidMkWriter(object):
                 if isinstance(v, list):
                     self.WriteList(v, k)
                 else:
-                    self.WriteLn("%s := %s" % (k, make.QuoteIfNecessary(v)))
+                    self.WriteLn("{} := {}".format(k, make.QuoteIfNecessary(v)))
             self.WriteLn("")
 
         # Add to the set of targets which represent the gyp 'all' target. We use the
@@ -928,7 +927,7 @@ class AndroidMkWriter(object):
         if self.target != self.android_module and write_alias_target:
             self.WriteLn("# Alias gyp target name.")
             self.WriteLn(".PHONY: %s" % self.target)
-            self.WriteLn("%s: %s" % (self.target, self.android_module))
+            self.WriteLn(f"{self.target}: {self.android_module}")
             self.WriteLn("")
 
         # Add the command to trigger build of the target type depending
@@ -975,25 +974,25 @@ class AndroidMkWriter(object):
     ):
         """Write a variable definition that is a list of values.
 
-    E.g. WriteList(['a','b'], 'foo', prefix='blah') writes out
-         foo = blaha blahb
-    but in a pretty-printed style.
-    """
+        E.g. WriteList(['a','b'], 'foo', prefix='blah') writes out
+             foo = blaha blahb
+        but in a pretty-printed style.
+        """
         values = ""
         if value_list:
             value_list = [quoter(prefix + value) for value in value_list]
             if local_pathify:
                 value_list = [self.LocalPathify(value) for value in value_list]
             values = " \\\n\t" + " \\\n\t".join(value_list)
-        self.fp.write("%s :=%s\n\n" % (variable, values))
+        self.fp.write(f"{variable} :={values}\n\n")
 
     def WriteLn(self, text=""):
         self.fp.write(text + "\n")
 
     def LocalPathify(self, path):
         """Convert a subdirectory-relative path into a normalized path which starts
-    with the make variable $(LOCAL_PATH) (i.e. the top of the project tree).
-    Absolute paths, or paths that contain variables, are just normalized."""
+        with the make variable $(LOCAL_PATH) (i.e. the top of the project tree).
+        Absolute paths, or paths that contain variables, are just normalized."""
         if "$(" in path or os.path.isabs(path):
             # path is not a file in the project tree in this case, but calling
             # normpath is still important for trimming trailing slashes.
@@ -1006,7 +1005,7 @@ class AndroidMkWriter(object):
         # so we don't look for a slash.
         assert local_path.startswith(
             "$(LOCAL_PATH)"
-        ), "Path %s attempts to escape from gyp path %s !)" % (path, self.path)
+        ), f"Path {path} attempts to escape from gyp path {self.path} !)"
         return local_path
 
     def ExpandInputRoot(self, template, expansion, dirname):
