@@ -8,7 +8,6 @@
 These functions are executed via gyp-mac-tool when using the Makefile generator.
 """
 
-from __future__ import print_function
 
 import fcntl
 import fnmatch
@@ -23,8 +22,6 @@ import subprocess
 import sys
 import tempfile
 
-PY3 = bytes != str
-
 
 def main(args):
     executor = MacTool()
@@ -33,7 +30,7 @@ def main(args):
         sys.exit(exit_code)
 
 
-class MacTool(object):
+class MacTool:
     """This class performs all the Mac tooling steps. The methods can either be
   executed directly, or dispatched from an argument list."""
 
@@ -179,7 +176,7 @@ class MacTool(object):
     def ExecCopyInfoPlist(self, source, dest, convert_to_binary, *keys):
         """Copies the |source| Info.plist to the destination directory |dest|."""
         # Read the source Info.plist into memory.
-        with open(source, "r") as fd:
+        with open(source) as fd:
             lines = fd.read()
 
         # Insert synthesized key/value pairs (e.g. BuildMachineOSBuild).
@@ -251,7 +248,7 @@ class MacTool(object):
 
         dest = os.path.join(os.path.dirname(info_plist), "PkgInfo")
         with open(dest, "w") as fp:
-            fp.write("%s%s" % (package_type, signature_code))
+            fp.write(f"{package_type}{signature_code}")
 
     def ExecFlock(self, lockfile, *cmd_list):
         """Emulates the most basic behavior of Linux's flock(1)."""
@@ -278,9 +275,7 @@ class MacTool(object):
         # epoch=0, e.g. 1970-1-1 or 1969-12-31 depending on timezone.
         env["ZERO_AR_DATE"] = "1"
         libtoolout = subprocess.Popen(cmd_list, stderr=subprocess.PIPE, env=env)
-        _, err = libtoolout.communicate()
-        if PY3:
-            err = err.decode("utf-8")
+        err = libtoolout.communicate()[1].decode("utf-8")
         for line in err.splitlines():
             if not libtool_re.match(line) and not libtool_re5.match(line):
                 print(line, file=sys.stderr)
@@ -540,7 +535,7 @@ class MacTool(object):
                 "application-identifier", ""
             )
             for team_identifier in profile_data.get("TeamIdentifier", []):
-                app_id = "%s.%s" % (team_identifier, bundle_identifier)
+                app_id = f"{team_identifier}.{bundle_identifier}"
                 if fnmatch.fnmatch(app_id, app_id_pattern):
                     valid_provisioning_profiles[app_id_pattern] = (
                         profile_path,
