@@ -5,7 +5,6 @@
 
 """gyptest.py -- test runner for GYP tests."""
 
-from __future__ import print_function
 
 import argparse
 import os
@@ -153,7 +152,7 @@ def print_configuration_info():
         sys.path.append(os.path.abspath("test/lib"))
         import TestMac
 
-        print("  Mac %s %s" % (platform.mac_ver()[0], platform.mac_ver()[2]))
+        print("  Mac {} {}".format(platform.mac_ver()[0], platform.mac_ver()[2]))
         print("  Xcode %s" % TestMac.Xcode.Version())
     elif sys.platform == "win32":
         sys.path.append(os.path.abspath("pylib"))
@@ -168,7 +167,7 @@ def print_configuration_info():
     print()
 
 
-class Runner(object):
+class Runner:
     def __init__(self, formats, tests, gyp_options, verbose):
         self.formats = formats
         self.tests = tests
@@ -217,10 +216,10 @@ class Runner(object):
             res = "skipped"
         elif proc.returncode:
             res = "failed"
-            self.failures.append("(%s) %s" % (test, fmt))
+            self.failures.append(f"({test}) {fmt}")
         else:
             res = "passed"
-        res_msg = " %s %.3fs" % (res, took)
+        res_msg = f" {res} {took:.3f}s"
         self.print_(res_msg)
 
         if (
