@@ -8,7 +8,6 @@
 generate input suitable for graphviz to render a dependency graph of
 targets."""
 
-from __future__ import print_function
 
 import collections
 import json
@@ -66,7 +65,7 @@ def WriteGraph(edges):
             target = targets[0]
             build_file, target_name, toolset = ParseTarget(target)
             print(
-                '  "%s" [shape=box, label="%s\\n%s"]' % (target, filename, target_name)
+                f'  "{target}" [shape=box, label="{filename}\\n{target_name}"]'
             )
         else:
             # Group multiple nodes together in a subgraph.
@@ -74,14 +73,14 @@ def WriteGraph(edges):
             print('    label = "%s"' % filename)
             for target in targets:
                 build_file, target_name, toolset = ParseTarget(target)
-                print('    "%s" [label="%s"]' % (target, target_name))
+                print(f'    "{target}" [label="{target_name}"]')
             print("  }")
 
     # Now that we've placed all the nodes within subgraphs, output all
     # the edges between nodes.
     for src, dsts in edges.items():
         for dst in dsts:
-            print('  "%s" -> "%s"' % (src, dst))
+            print(f'  "{src}" -> "{dst}"')
 
     print("}")
 
